+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## Unreleased
 
+- feat: Re-export Profiler and useProfiler from @sentry/react #1372
+- fix(performance): Handle edge cases in React Navigation routing instrumentation. #1365
 - build(android): Bump sentry-android to 4.3.0 #1373
 - build(devtools): Bump @sentry/wizard to 1.2.2 #1383
+- build(js): Bump sentry-javascript dependencies to 6.2.1 #1384
+- feat(performance): Option to set route change timeout in routing instrumentation #1370
 
 ## 2.2.2